+++ semana14-case/src/hooks/useRequestData.js
@@ -8,12 +8,11 @@ const useRequestData = (initialState, url) => {
         axios.get(url)
 
         .then((res) => {
-            console.log(res.data)
             setData(res.data)
         })
     
-        .catch((err) => {
-            console.log(err)
+        .catch(() => {
+            alert("Ocorreu um erro, tente novamente!")
         })
     }, [url])
