--- semana14-case/src/service/request.js
@@ -1,12 +1,0 @@
-// import axios from "axios"
-
-// export const getTarot = () => {
-//     axios.get("tarot.json")
-
-//     .then((res) => {
-//         console.log(res.data)
-//     })
-//     .catch((err) => {
-//         console.log(err)
-//     })
-// }
